git: Embed bundled Git binaries
When Omnibus conducts a zero-downtime upgrade of GitLab, it swaps Git
binaries on the system with new versions as required. This can cause
problems if Gitaly is actively invoking `git`, as symlinks are now
broken. https://gitlab.com/gitlab-org/gitaly/-/issues/5927 describes
the issue in more detail.

To address this, we should embed Git binaries within the `gitaly`
binary so updates are atomic.

Modify packedBinariesFS to include the Git binaries we build as part of
the `build-bundled-git` Makefile target. This target is executed when
Gitaly is configured to use a "bundled" Git installation. This is the
default for production, and subsequent commits will make this the
default for testing too.

Modify the Makefile to declare the newly packed binaries, and ensure
that Git is compiled before Gitaly. Retain the `install-bundled-git`
commands as they are required downstream, but will essentially become a
no-op.

For more information on Git execution environments, see the
documentation at doc/git-execution-environments.md.

Changelog: added