test: Default to using bundled Git
Modify testcfg.Build() to ensure we use the bundled Git environment by
default. Add a warning if the bundled Git environment was not compiled
with `make build-bundled-git` prior to running tests. This can occur if
the user runs tests via an IDE instead of `make test ...`, as that will
bypass the dependency to build Git.

Compiling Git adds some overhead to the initial test being run (about
20 seconds on my machine), but the compiled binaries remain in the
_build/bin directory for subsequent tests to use. This is true for CI
as well.