Add a friendly interface to access a storage's metadata KV
WAL's partition manager allows the caller to access a KV DB of a
partition. It's done by initializing a transaction against a partition.
The partition manager reserves partition ID 1 for storage-wide
metadata. Although the caller could start a transaction against that
pseudo partition, it's unintuitive to do so.

This commit adds a more friendly interface for such access.