Use original binary's mode for the new binary
`replace_buildid` replaces the temporary build ID of a Gitaly
binary with the final one. It currently uses the SharedExecutable
permissions for the new binary. Use the original permissions instead.
After all, we want to just replace the build ID in the binary and our
intention is not to change the permissions.