Remove production usage of PublicDir
Outside of tests, PublicDirs permission is used in two locations:

1. Quarantine directories are marked PublicDir.
2. `custom_hooks/` directories are marked PublicDir.

Quarantines are internal runtime state of Gitaly and don't need to
be readable by others. Custom hooks directories in the repositories
also shouldn't need to be accessible by anything but Gitaly.

Replace the usage of PublicDir with the standard storage permissions.