Introduce package for Posix permission bits
We're currently defining the modes we need using octals. These can
be difficult to read unless the reader is familiar with the octals
from before.

While all directories and files should be using the modes defined
by the storage layer in the `mode` package, there are some special
cases left that may still want to define their own modes. For example,
log files might need to be readable by other users than Gitaly's user
for them to be collected. Tests may for convenience want to use more
relaxed permissions.

Define a package to store Posix permission bit constants. This makes
the definitions easier to read and we can reuse the permission bits
elsewhere if we ever need to define other modes than the ones
supported by the storage layer.