Move ReadOnlyDirectory mode to snapshot package
Shared snapshots are made read-only by dropping the write permissions
from the directoris in the snapshot. The ReadOnlyDirectory mode is only
used for this purpose. Move it to the snapshot package as it's rather
an implementation detail of that package and not a general mode that
should be used by anything else in Gitaly.