Generalize the WALed permission field to mode
The WAL field hosting directory permissions is currently specifically
for permissions. Generalize the field to mode to better match the
semantics used elsewhere as well. This would also technically allow
us to now log non-permission bits in the mode, for example the sticky
bit, but we don't need to do so for now.