raft: Implement a deterministic ring-based replica placement strategy
This commit implements a deterministic replica placement. This strategy
appoints a set of replica groups for a particular storage. Data of a
storage are replicated to some certain amount of storages (defined by
its replication factor) next to it in the ring. The order of the
storage is determined by the storage ID, allocated by the metadata Raft
group.

For example:
- A cluster of 3 nodes has the following replica placement:
{
  storage-1 => [storage-2, storage-3],
  storage-2 => [storage-3, storage-1],
  storage-3 => [storage-1, storage-2],
}

- A cluster of 5 nodes has the following replica placement:
{
  storage-1 => [storage-2, storage-3],
  storage-2 => [storage-3, storage-4],
  storage-3 => [storage-4, storage-5],
  storage-4 => [storage-5, storage-1],
  storage-5 => [storage-1, storage-2],
}

- If storage-4 is decommissioned:
{
  storage-1 => [storage-2, storage-3],
  storage-2 => [storage-3, storage-5],
  storage-3 => [storage-5, storage-1],
  storage-5 => [storage-1, storage-2],
}

The strategy takes the physical storage residence into account. It
means two or more storages of a same node don't replicate to each
other. Although we limit one storage per node at this point, this
coverage saves us some headaches later.