git: Add code to parse reftable tables
We generally don't have to parse reftable's as this is handled entirely
by Git. But for testing the transaction manager, it is essential that we
test the integrity of files being added to the log. This is done by
simply looking at the refs present in the file. This is simple since the
files contain regular text.

Reftables however are a binary format, as such, we need to parse the
binary format to be able to understand the refs included in the file.

Let's add code to do this, currently the code is very limited to our use
case:
1. It only parses ref blocks and not other blocks.
2. It doesn't perform CR32 validation of the data since it is only
needed for reading the refs.
3. It is only allowed to run in tests. For all production code, we
should rely on Git and its implementation.