gitaly-ssh: Restore original working directory
A subsequent commit will execute `go env GOMOD` within tests to
determine the root directory of the Gitaly repository. This command
relies on the current working directly being within the Gitaly source
tree.

The gitaly-ssh command gets compiled and invoked by a test, and it
internally will change the working directory.

Modify the gitaly-ssh command so that it restores the previous working
directory after executing.