storagemgr: Filter transaction based on relative path
Now that we begin a transaction on a specific partition ID, by default
we want all repositories included in that transaction. So instead of the
relative path being overridden by the partition ID, we allow both, where
the transaction is started on a specific partition, and the relative
paths control which repositories are included in the snapshot.