Extract NewSubprocessLogger for hooks and lfs-smudge
lfs-smudge and hooks both contain duplicated logger setup logic.
Both log to a file. Extract the common functionality into a new
function so it's easier to modify the log setup in both commands.

`gitaly-lfs-smudge` had an issue where setting up the logger could
lead to panic as it would defer the `.Close()` call even if the
log setup fails and the returned closer is nil. We fix that while
at it. Otherwise the behavior should remain unchanged.