Add a test covering read-only snapshot removal after failure
Gitaly strips write permissions from directories in shared snapshots
as they are intended to be read-only. Removing the permissions ensures
nothing accidentally writes into them. PartitionManager's logic to
clean up the staging directory on start up isn't currently accounting
for the missing write permissions. This leads to Gitaly failing to
start as it can't clean up the staging directory.

Add a test case covering the scenario.

While at it, defer a missing close call on the dbMgr in the test.