Fix racy write in ServerInfo
ServerInfo is checking whether the storage is readable and writable.
The check is performed by statting the storage directory and writing
a test file in it. This has a number of issues:

1. If Gitaly doesn't exit cleanly, the test file is left stale in
   the storage.
2. Multiple invocations of the RPC will do the same write on the
   same file in a racy manner. While this doesn't itself cause issues,
   it's a bit surprising.
3. Testing for writes by writing a single file into the storage is
   not really sufficient. There could be issues writing in
   subdirectoreis of the storage as well.
4. The readability check done just stats the storage directory. It
   doesn't also ensure anything about the subdirectories. More over,
   statting a directory is possible without having read permissions
   on it, so the check is not really sufficient to ensure the storage
   can be read.
5. The check is ignoring errors coming from stat that could actually
   be useful.

All in all, the read/write check doesn't really make sense to have in
the first place. It doesn't sufficiently detect issues, and Gitaly
should refuse to boot anyway if it can't access or write in the
storage.

For now, remove the racy write in the storage by replacing the check
with whether or not we have the write permission on the file. We
should later look into removing the unnecessary functionality in
the RPC.