raft: Persist full storage info after registration
Recently, after registration, only the storage ID is persisted in
the storage's metadata DB. The metadata Raft group stores more
information than just the storage ID such as replication_factor. As
a result, we need to persist full storage information in the storage's
metadata DB. This persisted info will be used to compare with the
upstream info so that the storage is able to apply any changes.