Standardize file mode in Gitaly
Gitaly had previously a bunch of permissions, and they lived in the
perm package. As we're moving to remove the perm package, replace
all usage of `perm.PrivateWriteOnceFile` with `mode.File` and remove
the constant. The mode can be passed as is and the relevant bits are
extracted out of it in the standard library functions.