Move mode's tests to external package
Testhelper package will needs modes in some of its functions. This
leads to an import cycle as mode package needs to import testhelper
to call testhelper.Run in its TestMain as enforced by our linter.

Move the mode package's tests to an external package to avoid the
cycle. We need to do the same also for permission package as its
tests also need to import testhelper which then again import mode
and itself through there.