storagemgr: Extract out pack-refs code for files backend
The `preparePackRefs` method is used to prepare and apply refs packing
during housekeeping. The method is specific to the files backend. Since
we want to introduce the reftable backend, let's extract out the code to
`preparePackRefsFiles`. This allows us to add reftable specific methods
to `preparePackRefs` in the upcoming patches.