Wire log server into Gitaly
Gitaly's log server receives logs from subprocesses over a Unix
socket and outputs them with all the other logs in stdout. This
makes log management easier when all logs are in a single location
and not spread through different files.

This commit wires the log server into Gitaly. Gitaly starts a log
server when it starts serving. If the feature flag is enabled,
the subprocesses receive configuration to connect to the log server
and write their logs there instead of files.