raft: Add replication factor to Raft config
This commit adds the replication factor to the Raft config. This config
defines the number of nodes where the data of a storage are replicated.
It includes the main storage. For example, ReplicationFactor=3 means
1 main storage + 2 replicated storages.

The replication factor is set at the storage level. All repositories
of a storage share the same behavior. Although our Raft architecture
makes it feasible to set the replication factor at the repository
level, we don't need that fine-grained level of control at the moment.
This decision is subject to change in the future.

This commit also adds a field to cluster.proto. At this point,
cluster-related protobufs are not used anywhere except in tests. So,
it's okay to change protobuf field number, for the sake of simplicity.