storagemgr: Accept a list of relative paths to start a transaction
For maximum flexibility to specify the list of repositories in the
snapshot, we change transaction manager to accept a list of relative
paths. This more closely resembles the snapshot interface, and will
eventually allow us to differentiate between no repos and all repos.

There are too many tests to fix in this one commit, so for now we add a
workaround to let the tests pass and the tests themselves will be
refactored in a followup commit.