server: Add ServerSignature RPC
Contributes to https://gitlab.com/gitlab-org/gitaly/-/issues/6220

**Problem**

We need to create a Gitaly RPC to fetch the key used to sign web commits
so gitlab can return the public key to the end users to verify the web
commits are signed by Gitaly.

**Solution**

Add ServerSignature RPC that responds with a `PublicKey`.

Changelog: added