Expose metrics on started and stopped partitions
We're currently not exposing metrics on how many partitions have
been started and stopped. This would be useful information to
determine how many partitions are active at a given time, and also
how often we're starting and stopping them to determine whether
requests are reusing the same partition or spawning new ones.