raft: Update a storage's info if detect configuration change
When a node starts, there's a chance that node's Raft config has been
updated. If the changes include NodeID and/or ReplicationFactor, the
cluster needs to update the replica groups of all impacted storages.

So, when a node starts, the Raft manager compares the latest persisted
storage info to the node's Raft config. If it detects a change, it sends
an update request to the metadata Raft group. That group updates the
replica groups according to the replica placement strategy.

However, the manager deliberately avoids persisting the new storage
info. This is because another node could update the replica groups
without this node's awareness. So, the Raft manager should not attempt
to keep that info up-to-date.

Instead, we'll implement a replicator responsible for replication
management. It maintains a loop that polls the latest replica topology
from upstream and applies to the current node. It treats
the persisted info as the latest applied state.