repository: Wait acknowledgement for CreateBundleFromRefList
CreateBundleFromRefList immediately starts streaming the data without
waiting for acknowledgement first which makes it more difficult to
retry write requests as the client may not be able to retry with
another Gitaly as it may no longer have the data it is streaming.

Also it becomes more tricky to understand the actual error returned
from server if connection gets closed during the data stream.

Client should wait for acknowledgement from server first.