repository: Send acknowledgement for CreateBundleFromRefList
CreateBundleFromRefList does not acknowledge the header message sent
as the first message in the stream to the client which can make it
harder to understand the problem if server returns error while client
already starts its data stream without the acknowledgement.

Therefore server should acknowledge the initial request with empty
response and also validate that the first request dooes not contain
patterns.