featureflag: Default enable use_empty_tree_in_attr_tree_config
In Git change 2386535511, we introduced a feature:
 "attr: read attributes from HEAD when bare repo".
This causes a performance degradation.
Current workaround is to get rid of the default behavior
in bare repos of reading from HEAD by setting attr.tree to empty tree.

A feature flag use_empty_tree_in_attr_tree_config is added to control
the workaround. And it was enabled in production on Jun 24th.

This commit is to flip OnByDefault to true.