raft: Refactor dbAccessor interface
This interface was introduced to restrict access to the underlying
key-value DB. It is actually a function that receives a context,
a read-only boolean, and a callback. The read-only boolean argument is
fairly easy to misuse, leading to two consequences:
- Undetected DB updates in supposed read-only operations.
- Non-read-only operations are submitted as a new transaction even if
  there is literally no change. As a result, a significant overhead is
  added to each reading operation.

This commit refactors dbAccessor by creating a new struct consisting of
two distinct read and write operations.