doc: Cleanup up observability docs
There are two observability related files; dashboards.md and
observability.md. The latter one seems to be a subset of what
is in dashboards.md apart from a small difference. Also the
naming of the two files is bit confusing because, the content
of observability.md is just about grafana dashboards while the
content of dashboards.md is about general observability such as
logs, metrics, dashboards. Moving the content of dashboards.md
to observability.md, and removing the dashboards.md will make
things simpler and less confusing.