git: Add `NewRCVersion` to create an RC version
The `git.NewVersion` function allows us to create a new git version.
This is mostly used to check if a particular feature works with a
specific git version. Currently, it doesn't allow us to set the 'rc'
value, eventhough the underlying struct holds a 'rc' field.

Let's create a new `NewRCVersion` function to create a RC version
similar to the `NewVersion` function.

While we're here, we cleanup some of the older checks and usecases
since now git 2.45.0 is the minimum version.