server: Don't error on clean shutdown
Gitaly shutdowns can be caused by a variety of reasons:

- A "zero-downtime" upgrade to a new Gitaly binary using the tableflip
  library.
- Receiving a termination signal from the OS.
- An unrecoverable error.

In the first two cases, we attempt a graceful shutdown where existing
connections are drained before Gitaly exits. The grace period is
configurable via config.toml and defaults to one minute if not set.
Graceful shutdowns can also fail if Gitaly's gRPC servers have not
finished work by the time the period expires.

Currently, we log an "unclean Gitaly shutdown" message regardless of
whether the shutdown was done cleanly. This is because we return an
error even for successful graceful stops.

Stop returning an error if the graceful stop was successful. Instead,
replace the useful error message with additional info-level logging.