ci: Allow nightly jobs to fail
Nightly jobs test Gitaly against the `master` and `next` branches in
Git. These aren't used in production, so failures shouldn't block Gitlay
deployments.

Promote `allow_failure: true` to the root of the job definition.
Explicitly set `allow_failure: false` when the job is run as part of a
scheduled pipeline.