raft: Refactor Raft manager's restarting tests
In the test suite of Raft manager, there are some tests that assert the
state of the manager after a restart. They sets some internal states of
manager objects to `nil`. That approach is prone to be broken if the
manager introduces some more states. This commit refactors those tests.
They now re-create manager objects using existing storage instead.