storagemgr: Simplify log consumer factory
Recently, the PartitionManager triggers a factory to create a log
consumer with corresponding arguments. The factory returns a log
consumer object and a clean up function. This commit simplifies the log
consumer factory so that it returns the log consumer object only. The
partition manager triggers the cleanup method directly from that
consumer instead.