raft: Integrate WAL and Raft manager via Raft log consumer
In this commit, we create a new Raft consumer and injects it to the WAL
partition manager via the log consumer interface. Then, this log
consumer is connected to the Raft replicator inside Raft manager.
Afterward, the Raft replicator is now able to monitor partition changes
(deletion/creation) from the tracking storages.