Remove file logging support from subprocess logger
gitaly-hooks and gitaly-lfs-smudge are now writing their logs over
a pipe to the main Gitaly process that outputs them with all the
other logs. Remove the now unneeded functionality related to logging
into files.

gitaly-lfs-smudge has a slight behavior change where it starts to
fail if the logging configuration is not set. In practice the logging
configuration would always be set so this shouldn't be an issue.