testhelper: Use FileInfo to determine file mode in RequireTarState
Previously RequireTarState was casting tar.Header.Mode to
fs.FileMode but these two integers do not have the same layout. They
have a different number of bits. This meant that the assertion was not
detecting mode types.

Soon we will be using RequireTarState to assert tars that have symlinks
and so having mode types will be required.