beginners_guide: Modify the instructions to run tests with praefect
Running tests with praefect requires a postgres instance to be running.
While most users use GDK to fullfill this dependency, you can also do so
without GDK by simply spawning a postgres docker container. We already
have instructions to do so. However, we fail to add in that we also
should set the appropriate `env` variables, let's update the doc.