repository: Return error from stream on EOF
As documented on the ClientStream interface:
SendMsg is generally called by generated code. On error, SendMsg aborts
the stream. If the error was generated by the client, the status is
returned directly; otherwise, io.EOF is returned and the status of
the stream may be discovered using RecvMsg.

There are many places where we use chunker to send data. But majority
of those are ServerStream which does not have the same issue.
Therefore, instead of implementing error handling directly in chunker,
we decided to only touch the places which are using chunker for client
side streaming and implement the error handling in those places.