Implement a check for whether transactions have been active
If transactions have been active in the past but are no longer
active, some repositories may be in bad states if the WAL wasn't
fully applied to them prior to disabling transactions. We'll soon
add a middleware to ensure all WAL is applied. This commit implements
a check that is used to determine whether the middleware should be
used when transactions are disabled.