storagemgr: Extract out ReftableState
When testing for reftable's references, we simply consider the
`LooseReferences` stated as all the refs. This worked well because we
didn't have to go around patching every test. But it was short-sighted.
Let's split the `ReferencesState` into `FilesBackendState` and
`ReftableBackendState`. This allows us to extend the
`ReftableBackendState` in the following commits to check for number of
tables and min, max update index.