storagemgr: Add description for each subtest
Since we have a bunch of tests in the storagemgr package, we split them
into individual files to help make it more maintainable and readable.

While there is a logical file based split of the files, it is not
possible to split them while testing. Because, at the end, we merge all
the tests together before running. Let's instead add a short description
to each of the subtests, that makes it easier to only run tests from a
certail file/logical division.