storagemgr: Add additional checks for reftables
When doing a repository state assertion, for the files backend, we check
the references belonging to loose files and those in packed-refs.

For reftables, we only check if all the references are present. So let's
modify our tests to check for the contents of each table in
'tables.list'. With this, we check each table's references and its min
and max update index.

Amend the tests to consider these fields too.