Configure all accesors as retryable
Gitaly is currently configuring only a few fetch related RPCs as
retryable in the Go clients. In reality every accessor should be
retryable. The configuration is likely done with Workhorse and
Shell in mind but there may also be other users of the Go clients
such as the search indexers. Enable retries on every accessor in the
Go clients.

It's unfortunate that the service configs are spread between the Go
client and the Ruby client, and keeping them in sync is currently a
manual task. Keeping the service configs in sync is something we
should tackle later.