raft: Fix missing trailing "/" in DB's prefixes
Recently, all Raft-related keys have been missing trailing "/" in
their prefixes. As a result, they look like follows:
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/clusterapplied_lsn"
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/clustercluster"
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/selfstorage"

This missing does not affect the activities of Raft's package because
the namespacing is an application logical concept. The underlying
badger KV DB doesn't care.

That said, it makes sense to fix this namespacing issue to make it
easier to debug later. Afterward, the keys look like follows:
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/cluster/applied_lsn"
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/cluster/cluster"
* "p/\x00\x00\x00\x00\x00\x00\x00\x01/kv/raft/self/storage"