git: Return `[]Reference` from `IterateRefs()`
The `IterateRefs()` is used to iterate over and return all the
references present inside a given reftable. Earlier, it was returning
the values as `git.ReferenceUpdates`. This made sense as the usecase was
to treat them as such. Now that we want to also use this function to
check and validate the current state of each table in the transaction
manager, let's modify it to return `[]Reference`. This change is also
more in line with what the function actually does, since the reftables
hold the current state of each reference.