catfile: Add options to `ObjectReader` to disable mailmap
Currently we enable mailmap within the `ObjectReader` behind a feature
flag, eventually this will be made default mode of operation. However in
RPC's like 'GetCommitSignatures', we want to obtain the object contents
without mailmap. This is because if we enable mailmap, the contents of
the object will have mapped entries and signature verification will fail
since the commit could've been signed on the unmapped entry.

To avoid this situation, we need an option to disable mailmap in
`ObjectReader`. Add the required configuration and also tests for the
same.