catfile: Add `objectReadersWithoutMailmap` to cache
The catfile cache system allows us to cache cat-file '--batch-command'
processes. This ensures low overhead when we need to use one. However
the '--mailmap' option must be set when the process is spawned. It would
be nice if Git supported swapping mailmap options
mid-process (https://gitlab.com/gitlab-org/git/-/issues/344). But for
now this means we need to maintain two process caches, one with mailmap
and one without mailmap. So let's do that.