commit: Use `ObjectReaderWithoutMailmap` in `GetCommitSignatures`
When a commit is signed, it is either signed by the user or by Gitaly
itself. In both the scenarios the commit author can be different from
the output of the 'git-cat-file(1)' used with 'mailmap'. This difference
can cause the verification of the signature to fail. Since signature
verification is important, let's skip using 'mailmap' in the
`GetCommitSignatures` RPC.