commit: Add 'Author' to `GetCommitSignatures`
The `GetCommitSignatures` RPC was modified to not use 'mailmap' with
'git-cat-file(1)' recently. With this it is also important to provide
the actual author information to the clients that the commit signature
was verified with. This way, clients can validate and check if their
reference author is the same or if they have the 'mailmapp-ed' value.