backchannel: Don't use pointer when embedding struct
The mockTransactionServer struct is used for backchannel tests, and
embeds the gitalypb.UnimplementedRefTransactionServer type.

Version v1.5.1 of the protoc-gen-go-grpc package used to generate
the various *.pb.go files adds a check in the generated
RegisterRefTransactionServer function which calls the testEmbedByValue()
method of the provided server. Since we embedded the type with a pointer
receiver, this method did not exist and caused a panic.

Fix this by embedding the type directly.