bundleuri: Introduce bundle generation manager
A bundle generation manager will be responsible for handling requests to
generate bundles for repositories. It will be in control of these
processes and be able to stop them in the event Gitaly is shutting down.

It also will control concurrency and how many bundle generation
processes are allowed to be running at one time.

This will be passed in and used to automatically generate bundles for
bundle uri.