raft: Quarantine flaky tests related to manager reset
In the test suite, some tests assert the behavior of the Raft manager
after a restart. In such tests, we reset the state of the manager and
re-create nodehost instances. Right after the prior node host is
closed, there is a chance another process/another test occupies its
ephemeral port. That makes the next nodehost fail to bind.

The fix is not trivial. This commit quarantines such tests temporarily.