git: Fix `ReftableTableNameRegex` to read hex
The regex `ReftableTableNameRegex` is used to match table names and then
parse the min and max index values of the table. However it considers
the values to be integers, which is incorrect, since they are hex
values.

So let's first fix the regex and also introduce `git.ParseReftableName`
to actually parse the table name instead of repeating the logic in
multiple places. Add some tests too.